requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=2.0.0,!=2.1.0  # Apache-2.0
debtcollector>=1.2.0  # Apache-2.0
monasca-statsd>=1.1.0 # Apache-2.0
requests>=2.14.2 # Apache-2.0
PyYAML>=3.10  # MIT
six>=1.9.0 # MIT
monasca-common>=1.4.0 # Apache-2.0
oslo.config>=4.6.0  # Apache-2.0
oslo.log>=3.30.0  # Apache-2.0
